Initial commit
This code is all copied from scooter-map-frontend, and then rewritten to use jsdocs for typing instead of typescript. This means the project can get away with having zero tools – I think there's a lot of value to an open-source project having the most minimal dev environtment possible because it lowers the barrier to contributing, and also reduces my own headaches in getting configuration right.